Add build command to crater, and make a module to create launch scripts
Create the services module, which contains logic to create launch scripts for
the bundles' run_command's and main run_command defined in a Hipaacrate file.
These scripts are what run contain the logic to run the apps, and are required
by phusion/baseimage to actually run any service. These scripts are copied by the
resultant Dockerfile once built.

Add the build command to crater, which builds a Dockerfile from the Hipaacrate file.